[teatime] handle changes in format of args to tick
Currently the code assumes that tick will only send one arg. This code
checks if that is an int and if it's not assume its time and a rand
value to add additional entropy. This will be undefined until the
majority of clients update then we can turn on adding rand to the ticks.