[reflector] add dateInitialized property to island
This is used by the log aggregator script to figure out if a session has been
counted before or not. If the script only had the session id, it
wouldn't know if the session is a session that has already been counted, or if it
is a resuming session that needs to be counted again. The
dateInitialized field allows us to count the sessions with the same
dateInitialized only once.