[reflector] create NOTICE function for logging significant events
Also add an automated test to check that reflector.js logs out a "process
start" event.

Lastly, add a flag for disabling the time string at the beginning of each log
line when running locally. The flag is --no-logtime. The automated test
uses this to make it easier to parse stdout into json.