[dev-deploy] get rid of dev vs prod deployments in the same GCP project
Instead of having both dev and prod versions deployed in the same GCP
project, we want to instead have a separate GCP environment for each (or
a separate project beyond dev and prod). This involves removing
dichotomous files such as reflectors/build-dispatcher-dev-image and
reflectors/build-dispatcher-prod-image, and replacing them with a
consolidated file reflectors/build-image which takes in a project name
as an argument.

Additionally we can generate yaml files using a template yaml file along
with a script that substitutes placeholder values with project-name,
for example reflectors/generate-lb-urlmap.sh will replace values in
reflectors/lb-urlmap-dev-template.yaml and spit out a new file based on
the project name ($GCP_PROJECT-lb-urlmap.yaml). This gets rid of the
need to maintain separate yaml files for each project. Just generate a
new one when you need to for your project.