[dev-deploy] client dynamically chooses backend
The client library can now dynamically choose which backend urls to use regarding the sign func and the reflectors.

Either specify the backend using the  "backend" query parameter, or by default the client will look at the hostname and decide what the backend urls should be based off of the fact that the GCP project id could be in the hostname.

Defaults to prod urls if the hostname doesn't contain "croquet.dev"