[url-restriction] workaround for codepen iframe issue
Codepen sends the sign request from witin an iframe that has a different
origin from what you would expect, so url restriction doesn't work
properly. To get around this, we can look for the link with rel="canonical", which has the url we care about in its href attribute.
If there is no "canonical" link, we can stick with using the page's href
as before.