[cicd] add special case for croquet.dev
The url for the sign server and reflector is automatically created based
on the project name. However, croquet-dev is a special case since the
url may not contain the project name (since we host files on just
croquet.dev as opposed to <PROJECT_NAME>.croquet.dev). This fix sets the
project name to "croquet-dev" if there is no project name in the url.