[dev-deploy] client gets backend urls off of hostname
Previously the client would get backend urls by slicing the GCP project name off of the hostname, and using that GCP project name to construct the backend urls. To simplify and make things more flexible (such as to support "api.croquet.dev" which doesn't have a project name) we now construct the backend urls off of the hostname directly.