[bug-fix] use prod backend when file:// url or localhost
This fixes a bug that was introduced when the client was changed so that it could more flexibly choose backend urls based off of the hostname. The previous fix didn't consider if there was no hostname (such as on a file:// url) or if the url was a localhost url. Now it will work as it would before, choosing the prod backend in both the case of localhost urls and file:// urls.