[dev-deploy] fix bug in controller when choosing backend urls
I previously changed the controller to select backend urls based off of
the hostname so it could select different backends based off of which
GCP project the page was being served from. This worked for sites being
served from one of our GCP projects, but it no longer was choosing the
prod backend for sites served outside of our domains. This commit fixes
that by once again making it default to prod backend if the site isn't
served by one of our dev/staging domains.